% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_aos_refresh.R
\name{use_aos_refresh}
\alias{use_aos_refresh}
\title{Refresh aos}
\usage{
use_aos_refresh()
}
\value{
Javascript code
}
\description{
Recalculate all offsets and positions of elements (called on window resize).
}
\details{
By default AOS is watching for DOM changes and if there are any new elements loaded asynchronously or when something is removed from DOM it calls refreshHard automatically. In browsers that don't support MutationObserver like IE you might need to call AOS.refreshHard() by yourself.
}
\examples{
use_aos_refresh()

}
