% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_aos_refresh.R
\name{use_aos_refresh_hard}
\alias{use_aos_refresh_hard}
\title{Refresh hard aos}
\usage{
use_aos_refresh_hard()
}
\value{
Javascript code
}
\description{
Reinit array with AOS elements and trigger refresh (called on DOM changes that are related to aos elements)
}
\examples{
use_aos_refresh_hard()

}
