% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_aos.R
\name{html_dependencies_aos}
\alias{html_dependencies_aos}
\title{aos dependencies}
\usage{
html_dependencies_aos(use_cdn = TRUE)
}
\arguments{
\item{use_cdn}{use CDN (by default) or use locally stored files.}
}
\description{
Get aos html dependencies.
}
